refactor: move find files operations to standalone files
